solace_assistant/requirements.txt
pvporcupine
pyaudio
pvrecorder
numpy
whisper
pyttsx3
google.genai